requirements (1).txt
streamlit==1.29.0
faiss-cpu==1.7.4
numpy==1.26.3
sentence-transformers==2.2.2
pypdf2==3.0.1
re
